Added Indices
Now it's easier to navigate the vault thanks to indices to organize all
my notes. Indices are basically a markdown file which links all the
notes related to that specific subject/topic directly. This indexis then
linked in the homepage, also yes I improved the homepage with a lot more
additional information.